+++ .gitignore
@@ -1,1 +1,2 @@
 venv/
+app/.env
