+++ app/.env
@@ -1,1 +1,1 @@
-DISCORD_TOKEN_F1F = YOUR_TOKEN
+DISCORD_TOKEN_F1F = Votre Token
