+++ data/Barem.json
@@ -1,0 +1,9 @@
+{
+    "firstCorrect":2.5,
+    "secondCorrect":1.5,
+    "thirdCorrect":0.5,
+    "first2Correct":4,
+    "last2Correct":2.5,
+    "1and3Correct":3,
+    "allCorrect":5
+}
