+++ data/Leaderbord.json
@@ -1,0 +1,6 @@
+{
+    "1256311918319112294": {
+        "Pseudo": "victoirrrr",
+        "Points": 5
+    }
+}
