+++ data/Results.json
@@ -1,0 +1,6 @@
+{
+    "1": "Lando Norris",
+    "2": "Oscar Piastri",
+    "3": "Nico Hulkenberg",
+    "Best Lap": "Oscar Piastri"
+}
