+++ data/Session.json
@@ -1,0 +1,8 @@
+{
+    "Round": "13",
+    "Country": "Belgium",
+    "Location": "Spa-Francorchamps",
+    "Session": "Q",
+    "Date": "26/07/2025,14:00:00",
+    "Saison": 2025
+}
