--- data/resultats_course.json
@@ -1,14 +1,0 @@
-[
-    [
-        1,
-        "L. Norris"
-    ],
-    [
-        2,
-        "O. Piastri"
-    ],
-    [
-        3,
-        "N. Hülkenberg"
-    ]
-]
