--- data/resultats_qualif.json
@@ -1,14 +1,0 @@
-[
-    [
-        1,
-        "M. Verstappen"
-    ],
-    [
-        2,
-        "O. Piastri"
-    ],
-    [
-        3,
-        "L. Norris"
-    ]
-]
